Create aggregated data for SCEG-HiC

Function to generate aggregated inputs for SCEG-HiC Aggregate_datasets.

Usage

Aggregate_datasets(
  object,
  k_neigh = 50,
  rna_assay = "RNA",
  atac_assay = "peaks",
  cellnames = NULL,
  atacbinary = TRUE,
  max_overlap = 0.8,
  reduction.name = NULL,
  size_factor_normalize = TRUE,
  seed = 123,
  verbose = TRUE
)

Arguments

object

Seurat object.

k_neigh

Number of cells to be aggregated per group.

rna_assay

Name of assay containing gene expression information.

atac_assay

Name of assay containing peak information.

cellnames

Name of one or more metadata columns to group the cells. Default is the current cell identities.

atacbinary

Logical,whether the aggregated scATAC-seq data need binary?

max_overlap

The maximum overlapping ratio of two groups.

reduction.name

The reduction name of extracting the cell coordinates used for aggregating.

size_factor_normalize

Logical, should accessibility values be normalized by size factor?

seed

Random seed.

verbose

Logical, should warning and info messages be printed?

Value

Aggregated Seurat object.