Annotate transcription start sites (TSS)

This function annotates transcription start sites based on the specified species and genome assembly version.

Usage

annotateTSS(species, genome)

Arguments

species

Character string specifying the species name. Supported values are "Homo sapiens" or "Mus musculus".

genome

Character string specifying the genome assembly. Supported values are "hg38", "hg19", "mm10", or "mm9".

Value

A data.frame containing TSS annotations filtered by species and genome, including chromosome, gene name, and TSS genomic loci information.

Examples

# TSS annotations for the hg38
tssdata <- annotateTSS("Homo sapiens", "hg38")
head(tssdata)
#>    chr TargetGene TargetGeneTSS
#> 1 chr1    DDX11L1         11874
#> 2 chr1     WASH7P         29370
#> 3 chr1  MIR6859-1         17436
#> 4 chr1  MIR6859-2         17436
#> 5 chr1  MIR6859-3         17436
#> 6 chr1  MIR6859-4         17436

# TSS annotations for the mm10
tssdata <- annotateTSS("Mus musculus", "mm10")
head(tssdata)
#>     chr    TargetGene TargetGeneTSS
#> 1 chr12        Zc3h14      98746968
#> 2 chr15         Troap      99074973
#> 3  chr3         Clca2     145099106
#> 4 chr16 1810013L24Rik       8830100
#> 5 chr17         Srsf7      80207305
#> 6 chr15        Fignl2     101054399