Calculate promoters, enhancers, and Hi-C matrix of genes

The function obtains enhancers and promoters of each gene, and calculates the bulk average Hi-C matrix of the contact between the gene and enhancers.

Usage

calculateHiCWeights(
  SCEGdata,
  species,
  genome,
  focus_gene,
  averHicPath,
  TSSwindow = 1000,
  upstream = 250000,
  downstream = 250000,
  verbose = TRUE
)

Arguments

SCEGdata

Preprocessed data for SCEG-HiC.

species

String, indicating the name of the species.It can be either "Homo sapiens"or "Mus musculus".

genome

String, indicating the name of the genome assembly.It can be either "hg38", "hg19", "mm10", or "mm9".

focus_gene

The focused genes.

averHicPath

Path to the bulk average Hi-C.

TSSwindow

Numeric specifying the window size (in base pairs) to pad around either side of each TSS, to fetch gene's promoters. Default is 1000 bp (so 2000 bp window is drawn around each TSS).

upstream

Numeric specifying the window size (in base pairs) to pad around upstream of each TSS, to fetch gene's enhancers. Default is 250 kb (so 500 kb window is drawn around each TSS).

downstream

Numeric specifying the window size (in base pairs) to pad around downstream of each TSS, to fetch gene's enhancers. Default is 250 kb (so 500 kb window is drawn around each TSS).

verbose

Logical, should warning and info messages be printed?

Value

A list with enhancers, promoters, and Hi-C matrices for each focus_gene.

Examples

data(multiomic_small)
SCEGdata <- process_data(multiomic_small, k_neigh = 5, max_overlap = 0.5)
#> Generating aggregated data
#> Aggregating cluster 0
#> Sample cells randomly.
#> have 11 samples
#> Aggregating cluster 1
#> Sample cells randomly.
#> have 11 samples
fpath <- system.file("extdata", package = "SCEGHiC")
gene <- c("TRABD2A", "GNLY", "MFSD6", "CTLA4", "LCLAT1", "NCK2", "GALM", "TMSB10", "ID2", "CXCR4")
weight <- calculateHiCWeights(SCEGdata, species = "Homo sapiens", genome = "hg38", focus_gene = gene, averHicPath = fpath)
#> Successfully obtained 10 TSS loci of genes from chromosome 2.
#> Calculate weight of  TRABD2A
#> Calculate weight of  GNLY
#> Calculate weight of  MFSD6
#> Calculate weight of  CXCR4
#> Calculate weight of  CTLA4
#> Calculate weight of  LCLAT1
#> Calculate weight of  NCK2
#> Calculate weight of  ID2
#> Calculate weight of  GALM
#> Calculate weight of  TMSB10
weight[["CTLA4"]]$promoters
#> character(0)
weight[["CTLA4"]]$enhancers
#> [1] "chr2_203623664_203623982" "chr2_203682042_203682957"
#> [3] "chr2_203730093_203731243" "chr2_203786816_203787404"
#> [5] "chr2_203830381_203830563" "chr2_203932737_203935679"
#> [7] "chr2_203945543_203945939" "chr2_203992800_203993979"
#> [9] "chr2_204109915_204110871"
head(weight[["CTLA4"]]$contact)
#>    element1                 element2        score
#>      <char>                   <char>        <num>
#> 1:    CTLA4 chr2_203623664_203623982 0.0010420455
#> 2:    CTLA4 chr2_203682042_203682957 0.0009676338
#> 3:    CTLA4 chr2_203730093_203731243 0.0008481914
#> 4:    CTLA4 chr2_203786816_203787404 0.0014686981
#> 5:    CTLA4 chr2_203830381_203830563 0.0030630646
#> 6:    CTLA4 chr2_203932737_203935679 0.0020004300