Estimate size factors dense matrix

Estimate size factors dense matrix

Usage

estimateSizeFactorsForDenseMatrix(
  counts,
  locfunc = median,
  round_exprs = TRUE,
  method = "mean-geometric-mean-total"
)

Arguments

counts

The matrix for the gene expression data, either read counts or FPKM values or transcript counts.

locfunc

The location function used to find the representive value.

round_exprs

A logic flag to determine whether or not the expression value should be rounded.

method

A character to specify the size factor calculation appraoches. It can be either "mean-geometric-mean-total" (default), "weighted-median", "median-geometric-mean", "median", "mode", "geometric-mean-total".