Gene network reconstruction using prior biological knowledge

This function reconstructs a gene (or regulatory element) network by integrating prior biological knowledge into the estimation of the precision matrix. Instead of applying a uniform penalty to all element pairs, this method assigns different penalty values based on prior knowledge regarding whether two elements should be connected. Such prior information can be derived or obtained from biological data, such as a normalized bulk average Hi-C matrix.

Usage

wglasso(S, sample, normhicMatrix, element, alpha = NULL)

Arguments

S

A symmetric p-by-p covariance matrix representing co-variation between elements.

sample

An integer indicating the number of samples used to estimate the covariance matrix.

normhicMatrix

A normalized average Hi-C contact matrix used as prior biological knowledge for penalization.

element

A character vector containing the row/column names (element IDs) of the Hi-C matrix.

alpha

Multiplicative scaling factor for the penalty parameter (alpha * rho).

Value

A symmetric matrix of partial correlation coefficients between elements.