Run SCEG-HiC's model

The model uses data preprocessed by SCEG-HiC, the bulk average Hi-C as the penalty matrix, calculates the partial correlation coefficient matrix between elements using the wglasso model, and selects the gene-peak.

Usage

Run_SCEG_HiC(
  SCEGdata,
  HiCWeights,
  method = "wglasso",
  focus_gene,
  normalizeMethod = "1",
  cutoff = NULL,
  verbose = TRUE
)

Arguments

SCEGdata

Preprocessed data for SCEG-HiC.

HiCWeights

Raw the bulk average Hi-C matrix.Run the result of code'calculateHiCWeights'.

method

The method of gene network reconstruction.Default is wglasso.

focus_gene

The focused genes.

normalizeMethod

The method of normalizing the average Hi-C.Default is "1".

'1',rank score to normalize the average Hi-C.

'2',min-max to normalize the average Hi-C.

'3',0-1 to normalize the average Hi-C.

'4',-log10 to normalize the average Hi-C.

cutoff

The threshold of gene-peak.Default is NULL.If "aggregate" is TRUE ,we recommend the cutoff of 0.01. If "aggregate" is FALSE ,we recommend the cutoff of 0.001.

verbose

Logical, should warning and info messages be printed?

Value

A data.frame of correlations for each gene-peak pair.

Examples

data(multiomic_small)
SCEGdata <- process_data(multiomic_small, k_neigh = 5, max_overlap = 0.5)
#> Generating aggregated data
#> Aggregating cluster 0
#> Sample cells randomly.
#> have 11 samples
#> Aggregating cluster 1
#> Sample cells randomly.
#> have 11 samples
fpath <- system.file("extdata", package = "SCEGHiC")
gene <- c("TRABD2A", "GNLY", "MFSD6", "CTLA4", "LCLAT1", "NCK2", "GALM", "TMSB10", "ID2", "CXCR4")
weight <- calculateHiCWeights(SCEGdata, species = "Homo sapiens", genome = "hg38", focus_gene = gene, averHicPath = fpath)
#> Successfully obtained 10 TSS loci of genes from chromosome 2.
#> Calculate weight of  TRABD2A
#> Calculate weight of  GNLY
#> Calculate weight of  MFSD6
#> Calculate weight of  CXCR4
#> Calculate weight of  CTLA4
#> Calculate weight of  LCLAT1
#> Calculate weight of  NCK2
#> Calculate weight of  ID2
#> Calculate weight of  GALM
#> Calculate weight of  TMSB10
results_SCEGHiC <- Run_SCEG_HiC(SCEGdata, weight, focus_gene = gene)
#> The predicted genes are 10 in total.
#> Run model of  TRABD2A
#> Run model of  GNLY
#> Run model of  MFSD6
#> Run model of  CXCR4
#> Run model of  CTLA4
#> Run model of  LCLAT1
#> Run model of  NCK2
#> Run model of  ID2
#> Run model of  GALM
#> Run model of  TMSB10
results_SCEGHiC[results_SCEGHiC$gene == "CTLA4", ]
#>                           gene                     peak       score
#> chr2_203623664_203623982 CTLA4 chr2_203623664_203623982  0.03623216
#> chr2_203730093_203731243 CTLA4 chr2_203730093_203731243  0.14814205
#> chr2_203786816_203787404 CTLA4 chr2_203786816_203787404  0.00000000
#> chr2_203932737_203935679 CTLA4 chr2_203932737_203935679 -0.10013637
#> chr2_203945543_203945939 CTLA4 chr2_203945543_203945939  0.00000000
#> chr2_203992800_203993979 CTLA4 chr2_203992800_203993979  0.43240254
#> chr2_204109915_204110871 CTLA4 chr2_204109915_204110871  0.00000000