Create SCEG-HiC preprocessed input data

Function to create preprocessed input data for SCEG-HiC.calculateHiCWeights and Run_SCEG_HiCtakes as input the generated data.This function can generate clustered data or non clustered data for a specific cell type.

Usage

process_data(
  object,
  aggregate = TRUE,
  celltype = NULL,
  rna_assay = "RNA",
  atac_assay = "peaks",
  cellnames = NULL,
  atacbinary = TRUE,
  k_neigh = 50,
  max_overlap = 0.8,
  size_factor_normalize = TRUE,
  reduction.name = NULL,
  seed = 123,
  verbose = TRUE
)

Arguments

object

Seurat object.

aggregate

Logically, should the data be clustered ?If FALSE, celltype must be specified as one cell type.

celltype

Which cells to generate data.

rna_assay

Name of assay containing gene expression information.If aggregate is TRUE, name of assay containing raw gene expression information.If aggregate is FALSE, name of assay containing normalize gene expression information.

atac_assay

Name of assay containing peak information.If aggregate is TRUE, name of assay containing raw peak information. If aggregate is FALSE, name of assay containing normalize peak information.

cellnames

Name of one or more metadata columns to group the cells by. Default is the current cell identities.

atacbinary

Logical, whether the aggregated scATAC-seq data need binary?

k_neigh

Number of cells to be aggregated per group.

max_overlap

The maximum overlapping ratio of two groups.

size_factor_normalize

Logical, should accessibility values be normalized by size factor?

reduction.name

The reduction name of extracting the cell coordinates used for aggregating.

seed

Random seed.

verbose

Logical, should warning and info messages be printed?

Value

Preprocessed data for SCEG-HiC

Examples

data(multiomic_small)
# aggregated with paired scRNA-seq and scATAC-seq data
SCEGdata <- process_data(multiomic_small, k_neigh = 5, max_overlap = 0.5)
#> Generating aggregated data
#> Aggregating cluster 0
#> Sample cells randomly.
#> have 11 samples
#> Aggregating cluster 1
#> Sample cells randomly.
#> have 11 samples

# select "1" cells with paired scRNA-seq and scATAC-seq data
SCEGdata <- process_data(multiomic_small, aggregate = FALSE, celltype = "1")