Gene network reconstruction by integration of priori biological knowledge

The method, instead of using a single penalty, different penalty values were applied for element pairs based on a priori knowledge as to whether the two element should be connected. The a priori information can be calculated or retrieved from other biological data, e.g., the bulk average Hi-C matrix.

Usage

wglasso(S, sample, normhicMatrix, element)

Arguments

S

Covariance matrix:p by p matrix (symmetric)

sample

The number of samples.

normhicMatrix

A matrix with normalizing average Hi-C.

element

The row or col elements of the average Hi-C matrix.

Value

A matrix of partial correlation coefficients between elements.